+++ src/63createvertex.js
@@ -1,0 +1,99 @@
+/*
+//
+// CREATE VERTEX for AlaSQL
+// Date: 21.04.2015
+// (c) 2015, Andrey Gershun
+//
+*/
+
+yy.CreateVertex = function (params) { return yy.extend(this, params); }
+yy.CreateVertex.prototype.toString = function() {
+	var s = K('CREATE')+' '+K('VERTEX')+' ';
+	if(this.class) s += L(this.class)+' ';
+	if(this.sets) {
+		s += this.sets.toString();
+	} else if(this.content) {
+		s += this.content.toString();
+	} else if(this.select) {
+		s += this.select.toString();
+	}
+
+	return s;
+}
+
+yy.CreateVertex.prototype.toJavaScript = function(context, tableid, defcols) {
+		console.log('yy.CreateVertex.toJavaScript');
+	var s = 'this.queriesfn['+(this.queriesidx-1)+'](this.params,null,'+context+')';
+	return s;
+};
+// CREATE TABLE
+yy.CreateVertex.prototype.execute = function (databaseid,params,cb) {
+		console.log('yy.CreateVertex.execute');
+
+	if(this.sets) {
+		var obj = {};
+		if(this.sets.length > 0) {
+			this.sets.forEach(function(st){
+				console.log(st);
+			});
+		}
+
+	} else if(this.content) {
+
+	} else if(this.select) {
+
+	}
+
+
+	var res = 1;
+	if(cb) res = cb(res);
+	return res;
+};
+
+yy.CreateVertex.prototype.compile = function (databaseid) {
+	console.log('yy.CreateVertex.compile');
+
+	if(this.sets) {
+		var s = 'var a={};';
+		if(this.sets.length > 0) {
+			this.sets.forEach(function(st){
+				console.log(st);
+			});
+		}
+
+	} else if(this.content) {
+
+	} else if(this.select) {
+
+	}
+
+	return statement;
+};
+
+
+
+yy.CreateEdge = function (params) { return yy.extend(this, params); }
+yy.CreateEdge.prototype.toString = function() {
+//	console.log('here!');
+	var s = K('CREATE')+' '+K('EDGE')+' ';
+	if(this.class) s += L(this.class)+' ';
+	// SET
+	// CONTENT
+	// SELECT
+	return s;
+}
+
+yy.CreateEdge.prototype.toJavaScript = function(context, tableid, defcols) {
+	var s = 'this.queriesfn['+(this.queriesidx-1)+'](this.params,null,'+context+')';
+	return s;
+};
+
+// CREATE TABLE
+yy.CreateEdge.prototype.execute = function (databaseid,params,cb) {
+	var res = 1;
+	if(cb) res = cb(res);
+	return res;
+};
+
+yy.CreateEdge.prototype.compile = function (databaseid) {
+};