+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### 0.0.50 "Seoul" (21.04.2015 - 22.04.2015)
+* CREATE VERTEX and CREATE EDGE syntax
+* Fixed MIN and MAX functions and aggregators #93
+* Found UPDATE bug with column/columnid
+* Fixed bug with valueOf in comparision
+
 ### 0.0.49 "Beijing" (19.04.2015 - 21.04.2015)
 * CREATE CLASS
 * INSERT INTO class